add to history and detail fragments
add history cleaning, try to add saveState of DetilFragment

+++ FinalTask/app/src/main/java/com/example/finaltask/data/model/NumberResponse.kt
@@ -8,7 +8,7 @@ import com.squareup.moshi.JsonClass
 data class NumberResponse(
     @Json(name = "text") val text: String,
     @Json(name = "found") val found: Boolean,
-    @Json(name = "number") val number: Int,
+    @Json(name = "number") val number: Double,
     @Json(name = "type") val type: String,
     @Json(name = "date") val date: String?,
     @Json(name = "year") val year: String?
